Get MySQL username/password from environment
System.getenv(String) can be used to get environment variables.
These can be configured easily on individual machines, and in turn
allow tweaking and configuring without having to make changes to
the source code directly.

There is a fallback left in place, to default to the pre-existing
values, just in case someone was using that behavior.